Add a .mailmap to the repository so that git histories/logs know who has done what. For example:
Prior to adding a .mailmap
$ git shortlog -e -s
     1  Uri Guttman <uri AT PerlHunter.com>
     4  Uri Guttman <uri@cpan.org>
    33  Uri Guttman <uri@quad.(none)>

And after adding .mailmap
$ git shortlog -e -s
    38  Uri Guttman <uri@cpan.org>

It tells git to treat these possible differences in committers as
the same committer and normalizes a FirstName LastName <email>
for that committer.

This will greatly assist when looking for who has contributed to
the repository later (inclusions in Changes log and whatnot).