Remove work-arounds for undefined functions on old Perls.
Prior to Perl 5.6, several functions may or may not be defined
based on which version/platform you were on. To work around this,
these functions were tested for and defined in a BEGIN {} block.

Since this is no longer necessary, we can safely remove this block
given that Perl 5.6 is a requirement for installation.

Also, update the xt/author tests to no longer ignore these
work-arounds in the pod coverage tests.