Add regression tests for GLOB references.
In smoke testing all reverse dependencies of File::Slurp for a
proposed fix to the sysopen/sysread problem on Perl 5.30, an error
came up with a particular dist that makes use of GLOB references.

James Keenan then researched a good test case for us. Thank you,
James for going through all of that effort:
https://github.com/perhunter/slurp/pull/17#issuecomment-437174592

For his efforts, we now have another test case that we can add to
the current test suite to prevent us from committing any
regressions into this distribution.

We can now work on fixing up the other pull request with a bit
more faith that we won't break userland.