Refactor write_file() for 5.30 compliance.
Instead of syswrite, we use print();
Given that we can't override CORE::print(), we now SKIP the tests that
relied on overriding syswrite() to test failure points.
Fixup the prereqs a bit since we are also now using a few modules
in runtime that we previously only required for testing.
Now that we use print() there's no need to worry about the newlines on
windows since the layers provide that functionality.
Update the META.json since we have new prereqs.