Localize $^W to mitigate warnings from File::Temp.
File::Temp's tempfile warns when you ask for an unopened temporary file
to let you know that it might not be the safest option because the
library itself can't guarantee that another process won't pick that
filename prior to you opening the file yourself.

This presents a problem if the user runs their app with the -w flag
since we need to open the file ourselves to deal with file permissions
on the open process so we adhere to both the filesystem's potential ACL
settings as well as the original file's permission set.

This local $^W fix prevents the warning we know about from bothering the
end user as it's not a warning they should be concerned about as we are
dealing with it in File::Slurp.